backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
pypdf==5.4.0
python-docx==1.1.2
